refactor(vmm): lift the snapshot seam and device-model guard into mvm-vmm
SnapshotIO, the guarded load paths, the CannedIO double, and
assert_vsock_only_device_model now live in mvm-vmm, so a concrete backend
can implement the seam without a back-edge into mvm-runtime.

Narrower than planned, and deliberately so: pause_and_seal and
verify_and_resume stay in mvm-runtime with the store they depend on (the
on-disk layout, HMAC envelope and tenant-key encryption). Only the
guarded_* paths are store-free, and they are all a backend needs. Moving
the sealing helpers would have dragged the store down with them for no
caller.

The guard had briefly existed twice — the moved copy plus the original —
which for a claim-10 witness is worse than not having moved it. There is
one implementation again; mvm-runtime re-exports it at the old path.
Mutation surface re-pinned for the new location.
