+++ crates/mvm-vmm/src/lib.rs
@@ -17,6 +17,7 @@ pub mod driver;
 pub mod host;
 pub mod hvf_handoff;
 pub mod post_restore;
+pub mod snapshot;
 pub mod vmm;
 pub mod vsock_egress_bridge;
 pub mod vsock_transport;
